Fix filter and navigation
