+++ app/components/Navigation.tsx
@@ -1,0 +1,48 @@
+// app/components/Navigation.tsx
+import Link from "next/link";
+import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
+import { cookies } from "next/headers";
+
+export default async function Navigation() {
+  const supabase = createServerComponentClient({ cookies });
+  const {
+    data: { session },
+  } = await supabase.auth.getSession();
+
+  return (
+    <nav className="bg-white shadow">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <div className="flex justify-between h-16">
+          <div className="flex items-center">
+            {/* Logo/Home Link */}
+            <Link href="/" className="flex-shrink-0 flex items-center">
+              <h1 className="text-xl font-bold text-gray-900">DisñoJobs</h1>
+            </Link>
+          </div>
+
+          <div className="flex items-center space-x-4">
+            {session ? (
+              <>
+                <Link
+                  href="/jobs/create"
+                  className="inline-flex items-center px-4 py-2 border border-transparent 
+                           text-sm font-medium rounded-md text-white bg-blue-600 
+                           hover:bg-blue-700"
+                >
+                  Publicar Trabajo
+                </Link>
+                <Link href="/dashboard" className="text-gray-600 hover:text-gray-900">
+                  Dashboard
+                </Link>
+              </>
+            ) : (
+              <Link href="/auth" className="text-gray-600 hover:text-gray-900">
+                Iniciar Sesión
+              </Link>
+            )}
+          </div>
+        </div>
+      </div>
+    </nav>
+  );
+}
